ME3 file writing nearly complete.

+++ CoalescedConvert/Program.cs
@@ -9,6 +9,7 @@ namespace CoalescedConvert
 		private string _outputFileName;
 		private bool _encode;
 		private CoalescedFormat _format;
+		private bool _verbose;
 
 		static void Main(string[] args)
 		{
@@ -37,9 +38,11 @@ namespace CoalescedConvert
 			Console.WriteLine("CoelescedConvert.exe <coalesced_int.bin / coalesced_int.ini> [switches]");
 			Console.WriteLine();
 			Console.WriteLine("Switches");
-			Console.WriteLine("-h, --help  Show this help info.");
-			Console.WriteLine("-me12le     Use Mass Effect 1/2 Legendary Edition format.");
-			Console.WriteLine("-me3le      Use Mass Effect 3 Legendary Edition format.");
+			Console.WriteLine("-h, --help     Show this help info.");
+			Console.WriteLine("-v, --verbose  Show verbose logging.");
+			Console.WriteLine("-o, --out      Output file name (optional)");
+			Console.WriteLine("-me12le        Use Mass Effect 1/2 Legendary Edition format.");
+			Console.WriteLine("-me3le         Use Mass Effect 3 Legendary Edition format.");
 			Console.WriteLine();
 
 			return false;
@@ -47,12 +50,23 @@ namespace CoalescedConvert
 
 		private bool ProcessArguments(string[] args)
 		{
-			foreach (var arg in args)
+			for (int a = 0; a < args.Length; a++)
 			{
+				var arg = args[a];
 				var argLower = arg.ToLower();
+
 				if (argLower == "-h" || argLower == "--help")
 				{
 					return ShowUsage(null);
+				}
+				else if (argLower == "-v" || argLower == "--verbose")
+				{
+					_verbose = true;
+				}
+				else if (argLower == "-o" || argLower == "--out")
+				{
+					if (a + 1 >= args.Length) return ShowUsage($"Expected file name to follow '{arg}'.");
+					_outputFileName = args[++a];
 				}
 				else if (argLower == "-me12le")
 				{
@@ -77,13 +91,19 @@ namespace CoalescedConvert
 			var ext = Path.GetExtension(_inputFileName).ToLower();
 			if (ext == ".bin")
 			{
-				_outputFileName = Path.Combine(Path.GetDirectoryName(_inputFileName), string.Concat(Path.GetFileNameWithoutExtension(_inputFileName), ".ini"));
 				_encode = false;
+				if (string.IsNullOrEmpty(_outputFileName))
+				{
+					_outputFileName = Path.Combine(Path.GetDirectoryName(_inputFileName), string.Concat(Path.GetFileNameWithoutExtension(_inputFileName), ".ini"));
+				}
 			}
 			else if (ext == ".ini")
 			{
-				_outputFileName = Path.Combine(Path.GetDirectoryName(_inputFileName), string.Concat(Path.GetFileNameWithoutExtension(_inputFileName), ".bin"));
 				_encode = true;
+				if (string.IsNullOrEmpty(_outputFileName))
+				{
+					_outputFileName = Path.Combine(Path.GetDirectoryName(_inputFileName), string.Concat(Path.GetFileNameWithoutExtension(_inputFileName), ".bin"));
+				}
 			}
 			else
 			{
@@ -97,16 +117,20 @@ namespace CoalescedConvert
 		{
 			if (!ProcessArguments(args)) return 1;
 
+			Log.Verbose = _verbose;
+
 			if (_encode)
 			{
 				Console.WriteLine($"Converting INI:\n{_inputFileName}");
 				Console.WriteLine($"To BIN:\n{_outputFileName}");
 
+				if (!File.Exists(_inputFileName)) throw new FileNotFoundException($"The file '{_inputFileName}' could not be found.");
+
 				if (_format == CoalescedFormat.Unknown) _format = CoalescedFormatDetector.Detect(_outputFileName);
 				if (_format == CoalescedFormat.Unknown) throw new UnknownCoalescedFormatException();
 
 				var backupFileName = _outputFileName + ".bak";
-				if (!File.Exists(backupFileName))
+				if (File.Exists(_outputFileName) && !File.Exists(backupFileName))
 				{
 					Console.WriteLine($"Backup:\n{backupFileName}");
 					File.Copy(_outputFileName, backupFileName);
@@ -121,7 +145,11 @@ namespace CoalescedConvert
 						}
 						break;
 					case CoalescedFormat.MassEffect3LE:
-						throw new NotSupportedException();
+						{
+							var converter = new CoalescedConverterME3LE();
+							converter.Encode(_inputFileName, _outputFileName);
+						}
+						break;
 					default:
 						throw new UnknownCoalescedFormatException();
 				}
@@ -132,6 +160,8 @@ namespace CoalescedConvert
 			{
 				Console.WriteLine($"Converting BIN:\n{_inputFileName}");
 				Console.WriteLine($"To INI:\n{_outputFileName}");
+
+				if (!File.Exists(_inputFileName)) throw new FileNotFoundException($"The file '{_inputFileName}' could not be found.");
 
 				if (_format == CoalescedFormat.Unknown) _format = CoalescedFormatDetector.Detect(_inputFileName);
 				if (_format == CoalescedFormat.Unknown) throw new UnknownCoalescedFormatException();
